Begin making database access feel more like Haskell.
First off, passing around an IConnection to everything is just sloppy and
feels wrong.  So, hide all the connection specifics behind a Transaction
monad that everything can use.  This also requires creating a couple duplicates
of the standard HDBC functions to operate monadically, which in turn requires
qualifying imports for HDBC stuff.  But that should all be confined to the
DB module.

The next step is converting the DB module into more low-level code that provides
a way to construct high level operations without so much code duplication.