Add the editor, a program for graphically editing the database.
This is so I don't have to run sqlite by hand when I screw something up
and need to directly edit the database.  This program doesn't allow
editing everything but I typically don't require that.

It also allows some experimentation with esqueleto, which I think I am
going to start using for real in the next version of slog.