Add the beginnings of a full QSO adding program.
This is going to be a much more all-in-one program that allows for
looking up call signs, displaying the log as it changes, and showing
whether an entity has been worked already.

It will eventually replace all the graphical parts of qsoadd, which I
imagine can still be useful as a command line only sort of thing.