Replace existing database code with esqueleto-based stuff.
This just replaces the guts of the functions that were already there.  I
imagine there's probably better ways of interacting with the database
than what I was doing, but this is a start.  It also gets rid of the
monadic interface for the moment, which is likely going to have to come
back in order to avoid passing filenames all over the place and doing
the database migration if it doesn't exist.

But, this process has to start somewhere.  Here it is.