Convert all the DB functions to return a new data type.
Having only the QSO object sometimes and only the qsoid other times and
only the Confirmation object some other times is really inconvenient.
Now, all functions return a new type that includes all three of those
pieces of information.  This should cut down on the amount of database
calls required to piece everything together.

While I'm at it, add a couple of extra getQSOBy* functions that will be
convenient later and get rid of getQSOByID.  It was only needed for this
piecing together stuff I'm getting rid of.