Make the existing PreviousRow stuff more generic.
I'm about to start working on the code to display all QSOs on a
secondary tab.  It makes sense for that view to have all the same
columns as the previous QSOs view, so make the existing code generic
enough to support creating and populating both.

Note that I might want to add more columns later, probably taking some
ideas from what I've got in editor.