And then fix a whole lot of hlint warnings.
This doesn't deal with some spurious parse errors, uses of $ instead of
parens (I don't like using a lot of parens), nor some camel casing.
That last one is valid, but a lot of the ones it thinks I should camel
case match up with the names of things in ADIF or the radio control
language, and I want to keep those things identifiable.  So just ignore
them.