And then fix up the $ related warnings, too.
My objection to using () instead of $ is that I hate having a lot of
extra parens all over the place.  I think it makes the code harder to
read.  However, the problems hlint was complaining about actually just
move existing parens around to better suit the code, so I guess it's not
such a big deal.