Use smarter data types for contest mode.
Instead of the one ADT plus generating functions that confusingly return
two values plus storing both the function and the current value in the
program state, just create an existential type that stores everything.
Then there are some functions to generate values of this type for the
various contests, and I can get rid of the extra piece of program state.

I wanted to go farther and use typeclasses for this, but I couldn't make
it work.