Use rigctl to talk to rigctld, instead of connecting via network.
It's not really possible to know when we've gotten all the output we
were expecting since it could be one line or multiple lines.  Thus we
don't get an EOF, so the network reading functions never return.
Instead just use rigctl and rely on it to do whatever it does.