Support automatically setting the antenna based on frequency.
You've got to define an Antenna section in your .slog file and then
define a bunch of default_ lines for each band that you've got a default
antenna for.  It'd help if these were also listed under antennas=.  And
then if you have a default for everything else, you can define default=.

Basically I need to write documentation for this.

Also note that the antenna is set when focus leaves the frequency entry.
So if you need to override it, you need to set frequency first and then
change the antenna.  You can'd do the antenna first.