If the database does not exist, initialize it.
This is done with the new initDB function, which runs on startup of
every program and will verify the database exists with the right table
layout.

I also got rid of the stupid part in the editor program where it would
only look for the database in my home directory.  I've got a config
file, why am I not using it?