Change how antenna and mode combos are being populated.
I was completely defining them in code.  Now I am defining them in the
glade file as much as possible, but populating them in code.  This
allows untangling the whole mess a little bit to make it easier to
change the combos on the fly when QTH is changed.

I've got the beginnings of this here by adding QTH to the global program
state, and by doing the initial populating of the combo based on
whatever the default QTH is.