Add a new dialog to pick which QTH from the config file to use.
I originally wanted this to be a drop down menu, but those are pretty
much impossible to dynamically create.  Thus, I've got a new dialog that
comes out of a new configure menu option.  The code will keep track of
the currently configured QTH, but won't do anything with that
information.  I am saving database changes until the very end.

The code is now a big giant mess as far as I am concerned.  The next
steps are a lot of patches to refactor, split into multiple files, etc.