Add database columns for my callsign and my QTH.
This is how we're going to support multiple QTHes in one log.  Each
logged contact will get my callsign and my QTH added in as well.  That
should all work with this commit.  Then on the reporting side, I will
need to only upload one QTH's worth of contacts at a time to LoTW.  That
doesn't work yet.