If looking up a call sign fails, try looking it up in the DXCC table.
Basically, this is just trying harder to get useful information about
the call sign.  If hamqth doesn't know about the call in its databases,
we can instead ask to look up the call sign in the big DXCC table.  That
gives basic information like what entity the call's from, so we can
figure out if we need them or not.

This should eliminate a lot of the cases where "Nothing found for X"
gets printed.