Add a --no-rigctl command line option to slog.
This is needed for my radio, basically.  I need to be able to tell slog
to not run rigctl if I also plan on running wsjt or fldigi.  Those
programs can't use rigctl to communicate with my radio - they have to
talk to it directly.  Thus, I can't have both programs trying to access
the radio at the same time.