Only grab the split freq from rigctl if the radio supports it.
Basically, this makes my radio the special case and makes the code work
for grabbing the split frequency if it's supported by other radios.  I
feel like the rigctl interaction in the code is now a bit of a mess,
though.  There's going to have to be some reorganization soon.