Automatically update frequency display if rigctld is running.
Basically as you spin the VFO knob, we should periodically rigctl and
update the UI automatically.  This allows me to eliminate a couple
places where we were updating this part of the UI because the automatic
part will now handle it.  It also allows deleting some code I just
added.

This takes care of rigctl support as far as the TODO list is concerned,
for now.