Add a digital row to the checkmark grids.
I also moved all the checkmark grid code into its own file.  I think the
next thing to do is make it a little more obvious if a looked up
callsign is a new entity, mode, or band.  Having more checkmarks makes
it more difficult to tell now.

I also probably need to make it so it takes the current QTH into account
when displaying checkmarks, and then probably also add QTH to the
previous contacts views.  Displaying that you've contacted a station in
the view and then not displaying any checkmarks because you're in a
different QTH would be very confusing.