Make it easier to see if a contact is a new one.
I've added a grid of labels underneath the callsign entry that light up
when various conditions are met.  The intent is to make it faster to see
if a contact is an ATNO, or new for a band, or new for the grid.  If you
have the frequency entered, that information is taken into account.

This is not perfect, but I need to play with it for a while to decide
what might need to be changed.