Fix supporting multiple different contests in the same data structure.
I did this wrong to begin with.  In order to support the fact that there
can be a large number of different QSO record types, I need to add a
wrapper type around the CabrilloQSO type class and move where the
ExistentialQuantification is being used.  This actually isn't nearly as
bad as I thought it'd be.